+++ src/Stack/Config.hs
@@ -378,7 +378,7 @@ configFromConfigMonoid
          configScmInit = getFirst configMonoidScmInit
          configGhcOptions = configMonoidGhcOptions
          configSetupInfoLocations = configMonoidSetupInfoLocations
-         configPvpBounds = fromFirst PvpBoundsNone configMonoidPvpBounds
+         configPvpBounds = fromFirst (PvpBounds PvpBoundsNone False) configMonoidPvpBounds
          configModifyCodePage = fromFirst True configMonoidModifyCodePage
          configExplicitSetupDeps = configMonoidExplicitSetupDeps
          configRebuildGhcOptions = fromFirst False configMonoidRebuildGhcOptions
@@ -386,6 +386,7 @@ configFromConfigMonoid
          configAllowNewer = fromFirst False configMonoidAllowNewer
          configDefaultTemplate = getFirst configMonoidDefaultTemplate
          configDumpLogs = fromFirst DumpWarningLogs configMonoidDumpLogs
+         configSaveHackageCreds = fromFirst True configMonoidSaveHackageCreds
 
      configAllowDifferentUser <-
         case getFirst configMonoidAllowDifferentUser of